+++ yang/yang-model-util/src/main/java/org/opendaylight/yangtools/yang/model/util/type/BaseUint32Type.java
@@ -8,15 +8,16 @@
 package org.opendaylight.yangtools.yang.model.util.type;
 
 import org.eclipse.jdt.annotation.NonNull;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.opendaylight.yangtools.yang.model.api.type.Uint32TypeDefinition;
 import org.opendaylight.yangtools.yang.model.util.BaseTypes;
 
-final class BaseUint32Type extends AbstractRangeRestrictedBaseType<Uint32TypeDefinition, Long>
+final class BaseUint32Type extends AbstractRangeRestrictedBaseType<Uint32TypeDefinition, Uint32>
         implements Uint32TypeDefinition {
     static final @NonNull BaseUint32Type INSTANCE = new BaseUint32Type();
 
     private BaseUint32Type() {
-        super(BaseTypes.UINT32_QNAME, 0L, 4294967295L);
+        super(BaseTypes.UINT32_QNAME, Uint32.valueOf(0), Uint32.valueOf(4294967295L));
     }
 
     @Override
